Add files via upload
Project: Revealing the Formation Dynamics of Janus Polymer Particles: Insights from Experiments and Molecular Dynamics

by Miroslava Nedyalkova, Giovanni Russo,Philip Loche and Marco Lattuada

Contents of the repository: 
This release includes files to create initial structures for the MD model based on polystyrene particle, and SDS (sodium decile sulfate) and MMA (methyl metricate)  additives . 
The archive entitled “analysis.zip” contained the PMF results (data files and plots)  for the whole 10 ns simulations, and files ending in “-750” are for the last 2.5 ns of simulations (starting from frame 750). In the folder entitled "results.zip” – an archive with simulations can be found. 
File: plot_PMF.py" – a python script which can be used to generate the PMF graphs .
The scripts for plotting aspheristy, rdf, and for assessment the statistical distribution of the contains between the PS particle and the additive are available. 
The .itp files for the polysteine particle and MMA additive are available.